backend/requirements.txt
flask
flask-cors
pypdf2
pdfplumber
spacy
sentence-transformers
faiss-cpu
